+++ src/components/CharacterSheet.tsx
@@ -12,7 +12,7 @@ import {
     getPlayerCharacters,
     playerCharacters,
 } from 'services/playerService'
-import { useLocation, useParams } from 'solid-app-router'
+import { Link, useLocation, useParams } from 'solid-app-router'
 import {
     Component,
     createEffect,
@@ -24,6 +24,7 @@ import {
 } from 'solid-js'
 import { NonPlayerCharacter } from 'types/nonPlayerCharacter'
 import { PlayerCharacter } from 'types/playerCharacter'
+import WarningMessage from './WarningMessage'
 
 const CharacterSheet: Component = () => {
     const params = useParams()
@@ -106,40 +107,71 @@ const CharacterSheet: Component = () => {
     }
 
     return (
-        <div class="grid grid-cols-1 md:grid-cols-2 grid-rows-6">
-            <div class="grid grid-cols-1 md:grid-cols-2">
-                <label class="font-bold">Name:</label>
-                {selectedCharacter()?.prename} {selectedCharacter()?.surname}
-                <label class="font-bold">Clan:</label>
-                {selectedCharacter()?.clan}
-                <label class="font-bold">Erzeuger:</label>
-                {selectedCharacter()?.sire
-                    ? `${selectedCharacter()?.sire}`
-                    : '-'}
-                <label class="font-bold">Geschlecht:</label>
-                {translateSex(selectedCharacter()?.sex)}
-                <label class="font-bold">Generation:</label>
-                {selectedCharacter()?.generation ?? '?? '}th
-                <label class="font-bold">Verwandelt:</label>
-                {selectedCharacter()?.embraced ?? '-'}
-                <label class="font-bold">Sichtbares Alter:</label>
-                {selectedCharacter()?.apparentAge ?? '-'}
-                <label class="font-bold">Position:</label>
-                {selectedCharacter()?.position ?? '-'}
-                <label class="font-bold">Status:</label>
-                {translateStatus(selectedCharacter()?.characterStatus) ?? '-'}
+        <Show
+            when={selectedCharacter()}
+            fallback={
+                <div class="grid grid-cols-1 grid-rows-1">
+                    <WarningMessage message="Keine Daten vorhanden" />
+                </div>
+            }
+        >
+            <div class="grid grid-cols-1 md:grid-cols-2 grid-rows-6">
+                <div class="grid grid-cols-1 md:grid-cols-2">
+                    <label class="font-bold">Name:</label>
+                    {selectedCharacter()?.prename}{' '}
+                    {selectedCharacter()?.surname}
+                    <label class="font-bold">Clan:</label>
+                    {selectedCharacter()?.clan}
+                    <label class="font-bold">Erzeuger:</label>
+                    <Show
+                        when={selectedCharacter()?.sire_prename}
+                        fallback={'-'}
+                    >
+                        <Link
+                            href={`/npc/${
+                                selectedCharacter()?.sire_surname ?? ''
+                            }/${selectedCharacter()?.sire_prename}/`}
+                        >
+                            {selectedCharacter()?.sire_prename}{' '}
+                            {selectedCharacter()?.sire_surname}
+                        </Link>
+                    </Show>
+                    <label class="font-bold">Geschlecht:</label>
+                    {translateSex(selectedCharacter()?.sex)}
+                    <label class="font-bold">Generation:</label>
+                    {selectedCharacter()?.generation ?? '?? '}th
+                    <label class="font-bold">Verwandelt:</label>
+                    {selectedCharacter()?.embraced ?? '-'}
+                    <label class="font-bold">Sichtbares Alter:</label>
+                    {selectedCharacter()?.apparentAge ?? '-'}
+                    <label class="font-bold">Position:</label>
+                    {selectedCharacter()?.position ?? '-'}
+                    <label class="font-bold">Road:</label>
+                    <Show when={selectedCharacter()?.road} fallback={'-'}>
+                        <Link
+                            href={`/rules-info/road/${
+                                selectedCharacter()?.road
+                            }`}
+                        >
+                            {selectedCharacter()?.road}
+                        </Link>
+                    </Show>
+                    <label class="font-bold">Status:</label>
+                    {translateStatus(selectedCharacter()?.characterStatus) ??
+                        '-'}
+                </div>
+                <Show when={image()}>
+                    <img
+                        class="row-span-2 md:row-span-3 max-h-96 ml-0 md:ml-4 lg:ml-0 mt-4 md:mt-0"
+                        src={image()}
+                        loading="eager"
+                        alt={`Charakterbild von ${
+                            selectedCharacter()?.prename
+                        } ${selectedCharacter()?.surname}`}
+                    />
+                </Show>
             </div>
-            <Show when={image()}>
-                <img
-                    class="row-span-2 md:row-span-3 max-h-96 ml-0 md:ml-4 lg:ml-0 mt-4 md:mt-0"
-                    src={image()}
-                    loading="eager"
-                    alt={`Charakterbild von ${selectedCharacter()?.prename} ${
-                        selectedCharacter()?.surname
-                    }`}
-                />
-            </Show>
-        </div>
+        </Show>
     )
 }