+++ src/types/nonPlayerCharacterDirectus.ts
@@ -1,5 +1,9 @@
 import { NonPlayerCharacter } from './nonPlayerCharacter'
 
-export type NonPlayerCharacterDirectus = Omit<NonPlayerCharacter, 'clan'> & {
+export type NonPlayerCharacterDirectus = Omit<
+    NonPlayerCharacter,
+    'clan' | 'road'
+> & {
     clan: number
+    road: number
 }