+++ src/types/playerCharacter.ts
@@ -1,5 +1,6 @@
 import { CharacterStatus } from 'enums/characterStatus'
-import { ClanName } from 'enums/clan'
+import { ClanName } from 'enums/clanName'
+import { RoadName } from 'enums/roadName'
 import { Sex } from 'enums/sex'
 import { Status } from 'enums/status'
 import { DateTime } from 'luxon'
@@ -16,7 +17,9 @@ export interface PlayerCharacter {
     generation?: number
     position?: string
     prename: string
-    sire?: string
-    status: Status
     surname?: string
+    sire_prename?: string
+    sire_surname?: string
+    road?: RoadName | undefined
+    status: Status
 }