+++ src/types/playerCharacterDirectus.ts
@@ -1,5 +1,6 @@
 import { PlayerCharacter } from './playerCharacter'
 
-export type PlayerCharacterDirectus = Omit<PlayerCharacter, 'clan'> & {
+export type PlayerCharacterDirectus = Omit<PlayerCharacter, 'clan' | 'road'> & {
     clan: number
+    road: number
 }